Split marks and nodes into their own files
